python-fastapi/pyproject.toml
[project]
name = "fun-with-flags-demo-python"
version = "0.1.0"
description = "OpenFeature demo for Python (FastAPI) — end state of the Fun With Flags walkthrough."
requires-python = ">=3.10"
readme = "README.md"
authors = [
    { name = "Simon Schrottner", email = "simon.schrottner@gmail.com" },
]
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "openfeature-sdk>=0.7",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
